Function that checks if a binary tree is perfect

+++ 16-binary_tree_is_perfect.c
@@ -42,5 +42,5 @@ int binary_tree_is_perfect(const binary_tree_t *tree)
     height = binary_tree_height(tree);
     size = binary_tree_size(tree);
 
-    return (size == (1 << (height + 1)) - 1);
+return (size == ((size_t)1 << (height + 1)) - 1);
 }
